requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-filter==24.3
django-jazzmin==3.0.1
django-modeltranslation==0.19.12
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
Faker==36.1.1
gunicorn==23.0.0
inflection==0.5.1
Markdown==3.7
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
